feat: activity, bodyCondition 파일 분리 외
activity, bodyCondition 파일 분리
인스턴스 변수 main으로 이동
action 클로저 내부에서 printCurrentBodyCondition() 메서드 호출하도록 수정
